+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+## Fixed
+- Able to Deleted Project With Clusters In It from [mabhi](https://github.com/mabhi)
 ## [0.2.0] - 2023-01-27
 
 ## Fixed